renamed references to test.aws_mock_data
