+++ test/unit/aws/test_make_ami_public.py
@@ -7,7 +7,7 @@ import pytest
 from exasol.ds.sandbox.lib.aws_access.ami import Ami
 from exasol.ds.sandbox.lib.aws_access.aws_access import AwsAccess
 from exasol.ds.sandbox.lib.export_vm.run_make_ami_public import run_make_ami_public
-from test.aws_mock_data import get_ami_image_mock_data
+from test.aws.mock_data import get_ami_image_mock_data
 
 from test.mock_cast import mock_cast
